Moved reporting functions to separate file 🐾
