+++ src/utils/__tests__/colour-log.spec.ts
@@ -1,30 +1,15 @@
 import chalk from 'chalk'
-import { spaceLog } from 'space-log'
-
-import { Linter, type ReportSummary } from '@Types/lint'
 
 import colourLog from '../colour-log'
 
 jest.mock('chalk', () => ({
-  bgGreen: {
-    black: jest.fn().mockImplementation(text => text),
-  },
-  bgRed: {
-    black: jest.fn().mockImplementation(text => text),
-  },
-  bgYellow: {
-    black: jest.fn().mockImplementation(text => text),
-  },
   blue: jest.fn().mockImplementation(text => text),
   cyan: jest.fn().mockImplementation(text => text),
   dim: jest.fn().mockImplementation(text => text),
   magenta: jest.fn().mockImplementation(text => text),
   red: jest.fn().mockImplementation(text => text),
-  underline: jest.fn().mockImplementation(text => `_${text}_`),
   yellow: jest.fn().mockImplementation(text => text),
 }))
-
-jest.mock('space-log')
 
 jest.unmock('@Utils/colour-log')
 
@@ -32,16 +17,6 @@ describe('colourLog', () => {
 
   const mockedConsoleLog = jest.spyOn(console, 'log').mockImplementation(() => {})
   const mockedConsoleWarn = jest.spyOn(console, 'warn').mockImplementation(() => {})
-
-  const commonSummary: ReportSummary = {
-    deprecatedRules: [],
-    errorCount: 0,
-    fileCount: 1,
-    fixableErrorCount: 0,
-    fixableWarningCount: 0,
-    linter: Linter.ESLint,
-    warningCount: 0,
-  }
 
   describe('config', () => {
 
@@ -136,264 +111,6 @@ describe('colourLog', () => {
 
       expect(chalk.blue).toHaveBeenCalledOnceWith('Starting lint...')
       expect(mockedConsoleLog).toHaveBeenCalledOnceWith('Starting lint...')
-    })
-
-  })
-
-  describe('results', () => {
-
-    it('returns if there are no results', () => {
-      colourLog.results({
-        results: {},
-        summary: commonSummary,
-      })
-
-      expect(mockedConsoleLog).not.toHaveBeenCalled()
-    })
-
-    it('logs the results', () => {
-      const commonResult = {
-        message: 'Foo',
-        messageTheme: (str: string) => str,
-        position: '1:1',
-        positionTheme: (str: string) => str,
-        rule: 'bar',
-        ruleTheme: (str: string) => str,
-        severity: '  ×',
-        severityTheme: (str: string) => str,
-      }
-
-      colourLog.results({
-        results: {
-          'CONTRIBUTING.md': [commonResult],
-          'README.md': [commonResult, commonResult],
-        },
-        summary: commonSummary,
-      })
-
-      // Info
-      expect(chalk.blue).toHaveBeenCalledOnceWith('\nLogging eslint results:')
-      expect(mockedConsoleLog).toHaveBeenNthCalledWith(1, '\nLogging eslint results:')
-
-      // File 1
-      expect(mockedConsoleLog).toHaveBeenNthCalledWith(2)
-      expect(chalk.underline).toHaveBeenNthCalledWith(1, 'CONTRIBUTING.md')
-      expect(mockedConsoleLog).toHaveBeenNthCalledWith(3, '_CONTRIBUTING.md_')
-      expect(spaceLog).toHaveBeenNthCalledWith(1, {
-        columnKeys: ['severity', 'position', 'message', 'rule'],
-        spaceSize: 2,
-      }, [commonResult])
-
-      // File 2
-      expect(mockedConsoleLog).toHaveBeenNthCalledWith(4)
-      expect(chalk.underline).toHaveBeenNthCalledWith(2, 'README.md')
-      expect(mockedConsoleLog).toHaveBeenNthCalledWith(5, '_README.md_')
-      expect(spaceLog).toHaveBeenNthCalledWith(2, {
-        columnKeys: ['severity', 'position', 'message', 'rule'],
-        spaceSize: 2,
-      }, [commonResult, commonResult])
-
-      // Log Count
-      expect(mockedConsoleLog).toHaveBeenCalledTimes(5)
-      expect(chalk.underline).toHaveBeenCalledTimes(2)
-      expect(spaceLog).toHaveBeenCalledTimes(2)
-    })
-
-  })
-
-  describe('summary', () => {
-
-    let startTime: number
-
-    const expectResult = () => {
-      expect(chalk.cyan).toHaveBeenCalledWith('Finished eslint')
-      expect(chalk.yellow).toHaveBeenCalledWith('[1 file, 1000ms]')
-      expect(mockedConsoleLog).toHaveBeenCalledTimes(2)
-      expect(mockedConsoleLog).toHaveBeenNthCalledWith(1, '\nFinished eslint', '[1 file, 1000ms]')
-    }
-
-    beforeEach(() => {
-      jest.useFakeTimers().setSystemTime(1718971200)
-      startTime = new Date().getTime()
-      jest.advanceTimersByTime(1000)
-    })
-
-    it('logs the finished lint message along with the file count and duration (single file)', () => {
-      colourLog.summary(commonSummary, startTime)
-
-      expect(chalk.cyan).toHaveBeenCalledOnceWith('Finished eslint')
-      expect(chalk.yellow).toHaveBeenCalledOnceWith('[1 file, 1000ms]')
-      expect(mockedConsoleLog).toHaveBeenCalledOnceWith('\nFinished eslint', '[1 file, 1000ms]')
-    })
-
-    it('logs the finished lint message along with the file count and duration (multiple files)', () => {
-      colourLog.summary({
-        ...commonSummary,
-        fileCount: 7,
-      }, startTime)
-
-      expect(chalk.cyan).toHaveBeenCalledOnceWith('Finished eslint')
-      expect(chalk.yellow).toHaveBeenCalledOnceWith('[7 files, 1000ms]')
-      expect(mockedConsoleLog).toHaveBeenCalledOnceWith('\nFinished eslint', '[7 files, 1000ms]')
-    })
-
-    it('logs the error count in red (single error)', () => {
-      colourLog.summary({
-        ...commonSummary,
-        errorCount: 1,
-      }, startTime)
-
-      expectResult()
-      expect(chalk.red).toHaveBeenCalledWith('  1 error')
-      expect(mockedConsoleLog).toHaveBeenNthCalledWith(2, '  1 error')
-    })
-
-    it('logs the error count in red (multiple errors)', () => {
-      colourLog.summary({
-        ...commonSummary,
-        errorCount: 2,
-      }, startTime)
-
-      expectResult()
-      expect(chalk.red).toHaveBeenCalledWith('  2 errors')
-      expect(mockedConsoleLog).toHaveBeenNthCalledWith(2, '  2 errors')
-    })
-
-    it('logs the error count in red with the fixable error count in dim', () => {
-      colourLog.summary({
-        ...commonSummary,
-        errorCount: 3,
-        fixableErrorCount: 2,
-      }, startTime)
-
-      expectResult()
-      expect(chalk.red).toHaveBeenCalledWith('  3 errors')
-      expect(chalk.dim).toHaveBeenCalledWith(' (2 fixable)')
-      expect(mockedConsoleLog).toHaveBeenNthCalledWith(2, '  3 errors (2 fixable)')
-    })
-
-    it('logs the warning count in yellow (single warning)', () => {
-      colourLog.summary({
-        ...commonSummary,
-        warningCount: 1,
-      }, startTime)
-
-      expectResult()
-      expect(chalk.yellow).toHaveBeenCalledWith('  1 warning')
-      expect(mockedConsoleLog).toHaveBeenNthCalledWith(2, '  1 warning')
-    })
-
-    it('logs the warning count in yellow (multiple warnings)', () => {
-      colourLog.summary({
-        ...commonSummary,
-        warningCount: 5,
-      }, startTime)
-
-      expectResult()
-      expect(chalk.yellow).toHaveBeenCalledWith('  5 warnings')
-      expect(mockedConsoleLog).toHaveBeenNthCalledWith(2, '  5 warnings')
-    })
-
-    it('logs the warning count in yellow with the fixable warning count in dim', () => {
-      colourLog.summary({
-        ...commonSummary,
-        warningCount: 6,
-        fixableWarningCount: 3,
-      }, startTime)
-
-      expectResult()
-      expect(chalk.yellow).toHaveBeenCalledWith('  6 warnings')
-      expect(chalk.dim).toHaveBeenCalledWith(' (3 fixable)')
-      expect(mockedConsoleLog).toHaveBeenNthCalledWith(2, '  6 warnings (3 fixable)')
-    })
-
-    it('logs the deprecated rule count in magenta and the list in dim (single deprecation)', () => {
-      colourLog.summary({
-        ...commonSummary,
-        deprecatedRules: ['foo'],
-      }, startTime)
-
-      expectResult()
-      expect(chalk.magenta).toHaveBeenCalledWith('  1 deprecation')
-      expect(chalk.dim).toHaveBeenCalledWith(' [foo]')
-      expect(mockedConsoleLog).toHaveBeenNthCalledWith(2, '  1 deprecation [foo]')
-    })
-
-    it('logs the deprecated rule count in magenta and the list alphabetised in dim (multiple deprecations)', () => {
-      colourLog.summary({
-        ...commonSummary,
-        deprecatedRules: ['foo', 'bar', 'baz'],
-      }, startTime)
-
-      expectResult()
-      expect(chalk.magenta).toHaveBeenCalledWith('  3 deprecations')
-      expect(chalk.dim).toHaveBeenCalledWith(' [bar, baz, foo]')
-      expect(mockedConsoleLog).toHaveBeenNthCalledWith(2, '  3 deprecations [bar, baz, foo]')
-    })
-
-    it('logs everything together', () => {
-      colourLog.summary({
-        ...commonSummary,
-        deprecatedRules: ['foo', 'bar', 'baz'],
-        errorCount: 2,
-        fixableErrorCount: 1,
-        warningCount: 3,
-        fixableWarningCount: 2,
-      }, startTime)
-
-      expectResult()
-      expect(chalk.red).toHaveBeenCalledWith('  2 errors')
-      expect(chalk.dim).toHaveBeenCalledWith(' (1 fixable)')
-      expect(chalk.yellow).toHaveBeenCalledWith('  3 warnings')
-      expect(chalk.dim).toHaveBeenCalledWith(' (2 fixable)')
-      expect(chalk.magenta).toHaveBeenCalledWith('  3 deprecations')
-      expect(chalk.dim).toHaveBeenCalledWith(' [bar, baz, foo]')
-      expect(mockedConsoleLog).toHaveBeenNthCalledWith(2, '  2 errors (1 fixable)\n  3 warnings (2 fixable)\n  3 deprecations [bar, baz, foo]')
-    })
-
-  })
-
-  describe('summaryBlock', () => {
-
-    it('logs the error count in a red background', () => {
-      colourLog.summaryBlock({
-        ...commonSummary,
-        errorCount: 1,
-      })
-
-      expect(chalk.bgRed.black).toHaveBeenCalledOnceWith(' 1 ESLint Error ')
-      expect(mockedConsoleLog).toHaveBeenCalledOnceWith('\n🚨  1 ESLint Error ')
-    })
-
-    it('logs the warning count in a yellow background', () => {
-      colourLog.summaryBlock({
-        ...commonSummary,
-        warningCount: 1,
-      })
-
-      expect(chalk.bgYellow.black).toHaveBeenCalledOnceWith(' 1 ESLint Warning ')
-      expect(mockedConsoleLog).toHaveBeenCalledOnceWith('\n🚧  1 ESLint Warning ')
-    })
-
-    it('logs the both the error and warning counts if both are present', () => {
-      colourLog.summaryBlock({
-        ...commonSummary,
-        errorCount: 2,
-        warningCount: 3,
-      })
-
-      expect(chalk.bgRed.black).toHaveBeenCalledOnceWith(' 2 ESLint Errors ')
-      expect(chalk.bgYellow.black).toHaveBeenCalledOnceWith(' 3 ESLint Warnings ')
-      expect(mockedConsoleLog).toHaveBeenCalledTimes(2)
-      expect(mockedConsoleLog).toHaveBeenNthCalledWith(1, '\n🚨  2 ESLint Errors ')
-      expect(mockedConsoleLog).toHaveBeenNthCalledWith(2, '\n🚧  3 ESLint Warnings ')
-    })
-
-    it('logs a success message if there are no errors or warnings', () => {
-      colourLog.summaryBlock(commonSummary)
-
-      expect(chalk.bgGreen.black).toHaveBeenCalledOnceWith(' ESLint Success! ')
-      expect(mockedConsoleLog).toHaveBeenCalledOnceWith('\n✅  ESLint Success! ')
     })
 
   })
